Switched from Long to Integer in entities, added builders.

+++ src/main/java/ru/nfm/calendar/model/CalendarUserKey.java
@@ -8,6 +8,6 @@ import java.io.Serializable;
 @Embeddable
 @Data
 public class CalendarUserKey implements Serializable {
-    private Long userId;
-    private Long calendarId;
+    private Integer userId;
+    private Integer calendarId;
 }
